requirements.txt
playwright
pymupdf
ollama
python-telegram-bot
apscheduler
python-dotenv
pytest
pytest-asyncio
